backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
boto3==1.35.41
pyjwt==2.9.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-core==2.23.4
email-validator==2.2.0
